requirements.txt
absl-py==0.15.0
asgiref==3.5.2
astunparse==1.6.3
cachetools==5.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
dill==0.3.5.1
Django==3.2
django-colorfield==0.7.0
filelock==3.7.0
flatbuffers==1.12
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
huggingface-hub==0.0.12
idna==3.3
importlib-metadata==4.11.4
joblib==1.1.0
keras==2.8.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.7
numpy==1.19.5
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.53
sentencepiece==0.1.96
six==1.15.0
sqlparse==0.4.2
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-addons==0.17.0
tensorflow-estimator==2.5.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tokenizers==0.10.3
torch==1.11.0
tqdm==4.64.0
transformers==4.8.2
typeguard==2.13.3
typing-extensions==3.7.4.3
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.12.1
zipp==3.8.0
